backend/shared/requirements.txt
fastapi==0.115.8
pika==1.3.2
pydantic==2.10.6
pyjwt==2.10.1
SQLAlchemy==2.0.37
types-pika==1.2.0b1
